oci: take ExecCmd
